+++ test 1/Weight.swift
@@ -11,10 +11,10 @@ import SwiftData
 @Model
 class Weight : Identifiable {
     var id: Int
-    var weightvalue: Int = 150
+    var weightvalue: Float = 150
     var datetaken: Date = Date()
     
-    init(id: Int, weightvalue: Int, datetaken: Date) {
+    init(id: Int, weightvalue: Float, datetaken: Date) {
         self.id = id
         self.weightvalue = weightvalue
         self.datetaken = datetaken
